+++ packages/windie-sdk-js/cjs/index.js
@@ -17,7 +17,7 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
     for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.WindieAgentSession = exports.createWindieAgentSession = exports.createWindieAgentBackendTransport = exports.createAgentSession = exports.createAgentBackendTransport = exports.AgentSession = exports.createWindieLocalRuntimeProvider = exports.WindieClient = exports.WindieAgent = exports.WindieChatSession = void 0;
+exports.createAgentSession = exports.createAgentBackendTransport = exports.AgentSession = exports.createWindieLocalRuntimeProvider = exports.WindieClient = exports.WindieAgent = exports.WindieChatSession = void 0;
 __exportStar(require("./conversation/types.js"), exports);
 __exportStar(require("./conversation/events.js"), exports);
 __exportStar(require("./conversation/metadata.js"), exports);
@@ -60,7 +60,3 @@ var AgentSession_js_1 = require("./transport/AgentSession.js");
 Object.defineProperty(exports, "AgentSession", { enumerable: true, get: function () { return AgentSession_js_1.AgentSession; } });
 Object.defineProperty(exports, "createAgentBackendTransport", { enumerable: true, get: function () { return AgentSession_js_1.createAgentBackendTransport; } });
 Object.defineProperty(exports, "createAgentSession", { enumerable: true, get: function () { return AgentSession_js_1.createAgentSession; } });
-var WindieAgentSession_js_1 = require("./transport/WindieAgentSession.js");
-Object.defineProperty(exports, "createWindieAgentBackendTransport", { enumerable: true, get: function () { return WindieAgentSession_js_1.createWindieAgentBackendTransport; } });
-Object.defineProperty(exports, "createWindieAgentSession", { enumerable: true, get: function () { return WindieAgentSession_js_1.createWindieAgentSession; } });
-Object.defineProperty(exports, "WindieAgentSession", { enumerable: true, get: function () { return WindieAgentSession_js_1.WindieAgentSession; } });